Cast ONE_HOT_CLASSES to torch.float type

+++ mmsegmentation-clone/mmseg/datasets/order_pred.py
@@ -31,7 +31,7 @@ class OrderPredDataset(CustomDataset):
                '{c,b,a,d}')
     
     # the classes, but represented as one-hot tensors for loss calculation
-    ONE_HOT_CLASSES = torch.nn.functional.one_hot(torch.arange(len(CLASSES)))
+    ONE_HOT_CLASSES = torch.nn.functional.one_hot(torch.arange(len(CLASSES)), dtype=torch.float)
 
     PALETTE = None
